Setting 'viewport' to 'device-width'
This should be the ultimate—and apparently the modern standard—way to tell mobile devices that this is a mobile friendly site and to completely remove the after-tap delay. This unfortunately makes the width of the game all wonky on mobile, which is why we initially moved away from it... but, we'll just have to burn the bridges as we come to them in sequence, one at a time.